+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## `0.3.0`
+
+- Add language server client
+- Improve language configuration
+
 ## `0.2.1`
 
 - Add icon